make importUSDZTexturesFilePath const and move logic to initialize it to the initializer
